move client transaction code to log_comm.h; log_impl.h is server/sa mode only
