+++ src/base/error_manager.c
@@ -38,6 +38,7 @@
 #endif /* !SERVER_MODE */
 #include "memory_alloc.h"
 #include "message_catalog.h"
+#include "object_representation.h"
 #if !defined (WINDOWS)
 #include "release_string.h"
 #endif // not WINDOWS
