+++ src/communication/network_interface_cl.c
@@ -3121,107 +3121,6 @@ tran_server_2pc_prepare_global_tran (int gtrid)
 #endif /* !CS_MODE */
 }
 
-#if defined (ENABLE_UNUSED_FUNCTION)
-/*
- * tran_server_start_topop -
- *
- * return:
- *
- *   topop_lsa(in):
- *
- * NOTE:
- */
-int
-tran_server_start_topop (LOG_LSA * topop_lsa)
-{
-#if defined(CS_MODE)
-  int success = ER_FAILED;
-  int req_error;
-  OR_ALIGNED_BUF (OR_INT_SIZE + OR_LOG_LSA_ALIGNED_SIZE) a_reply;
-  char *reply;
-  char *ptr;
-
-  reply = OR_ALIGNED_BUF_START (a_reply);
-
-  req_error =
-    net_client_request (NET_SERVER_TM_SERVER_START_TOPOP, NULL, 0, reply, OR_ALIGNED_BUF_SIZE (a_reply), NULL, 0, NULL,
-			0);
-  if (!req_error)
-    {
-      ptr = or_unpack_int (reply, &success);
-      ptr = or_unpack_log_lsa (ptr, topop_lsa);
-    }
-  return success;
-#else /* CS_MODE */
-  int success = ER_FAILED;
-
-  THREAD_ENTRY *thread_p = enter_server ();
-
-  if (xtran_server_start_topop (thread_p, topop_lsa) != NO_ERROR)
-    {
-      success = ER_FAILED;
-    }
-  else
-    {
-      success = NO_ERROR;
-    }
-
-  exit_server (*thread_p);
-
-  return success;
-#endif /* !CS_MODE */
-}
-
-/*
- * tran_server_end_topop -
- *
- * return:
- *
- *   result(in):
- *   topop_lsa(in):
- *
- * NOTE:
- */
-TRAN_STATE
-tran_server_end_topop (LOG_RESULT_TOPOP result, LOG_LSA * topop_lsa)
-{
-#if defined(CS_MODE)
-  TRAN_STATE tran_state = TRAN_UNACTIVE_UNKNOWN;
-  int req_error, tran_state_int;
-  OR_ALIGNED_BUF (OR_INT_SIZE) a_request;
-  char *request;
-  OR_ALIGNED_BUF (OR_INT_SIZE + OR_LOG_LSA_ALIGNED_SIZE) a_reply;
-  char *reply;
-  char *ptr;
-
-  request = OR_ALIGNED_BUF_START (a_request);
-  reply = OR_ALIGNED_BUF_START (a_reply);
-
-  ptr = or_pack_int (request, (int) result);
-  req_error =
-    net_client_request (NET_SERVER_TM_SERVER_END_TOPOP, NULL, 0, reply, OR_ALIGNED_BUF_SIZE (a_reply), NULL, 0, NULL,
-			0);
-  if (!req_error)
-    {
-      ptr = or_unpack_int (reply, &tran_state_int);
-      tran_state = (TRAN_STATE) tran_state_int;
-      ptr = or_unpack_log_lsa (ptr, topop_lsa);
-    }
-  return tran_state;
-#else /* CS_MODE */
-  TRAN_STATE tran_state = TRAN_UNACTIVE_UNKNOWN;
-
-  THREAD_ENTRY *thread_p = enter_server ();
-
-  tran_state = xtran_server_end_topop (thread_p, result, topop_lsa);
-
-  exit_server (*thread_p);
-
-  return (tran_state);
-#endif /* !CS_MODE */
-}
-#endif
-
 /*
  * tran_server_savepoint -
  *
