+++ src/executables/util_common.c
@@ -34,8 +34,8 @@
 #include "utility.h"
 #include "util_func.h"
 #include "porting.h"
+#include "log_comm.h"
 #include "message_catalog.h"
-#include "log_impl.h"
 #include "mprec.h"
 #include "system_parameter.h"
 #include "environment_variable.h"
