+++ src/transaction/boot_cl.c
@@ -42,11 +42,13 @@
 
 #include <assert.h>
 
+#include "boot_cl.h"
+
 #include "porting.h"
 #if !defined(HPUX)
 #include "util_func.h"
 #endif /* !HPUX */
-#include "boot_cl.h"
+#include "connection_globals.h"
 #include "memory_alloc.h"
 #include "area_alloc.h"
 #include "storage_common.h"
@@ -62,6 +64,7 @@
 #endif /* !WINDOWS */
 #include "transaction_cl.h"
 #include "log_comm.h"
+#include "lzo/lzoconf.h"
 #include "server_interface.h"
 #include "release_string.h"
 #include "system_parameter.h"
