+++ src/transaction/locator_cl.c
@@ -27,6 +27,7 @@
 
 #include <stdio.h>
 #include <assert.h>
+#include <signal.h>
 
 #include "db.h"
 #include "environment_variable.h"
